requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.7.1

# Kafka
kafka-python==2.0.2

# Spark
pyspark==3.5.1

# Storage
duckdb==0.10.3

# Dashboard
streamlit==1.34.0
plotly==5.22.0
pandas==2.2.2

# LLM
anthropic==0.26.0

# Orchestration
apache-airflow==2.9.1

# Data quality
great-expectations==0.18.15

# Testing
pytest==8.2.0
pytest-cov==5.0.0

# Utilities
loguru==0.7.2
rich==13.7.1
typer==0.12.3